# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Wesley's Ambacht Catering API
  description: |
    # Wesley's Ambacht Catering API

    A comprehensive REST API for managing catering bookings, quotes, availability, and add-on services for Wesley's Ambacht, a premium Dutch catering company.

    ## Features

    - **Real-time Availability Management** - Live booking slots with automatic reservation
    - **Dynamic Quote Generation** - Intelligent pricing based on service tiers and guest counts
    - **Add-on Services** - Flexible service extensions with per-person and flat-rate pricing
    - **Multi-tier Service Categories** - Essential, Premium, and Luxury tiers for different events
    - **Real-time Updates** - WebSocket subscriptions for live booking and availability updates
    - **Comprehensive Logging** - Full request/response logging with performance metrics

    ## Authentication

    This API uses Supabase Row Level Security (RLS) with JWT tokens. Include your authorization token in the `Authorization` header:

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting

    - 100 requests per minute for general endpoints
    - 500 requests per minute for availability checks
    - 50 requests per minute for booking creation

    ## Error Handling

    All endpoints return structured error responses:

    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Guest count must be between 10 and 200",
        "details": {
...
        }
...

...
servers:
...
paths:
...
  /bookings/{id}:
    get:
      summary: Get booking details
      description: |
        Retrieve detailed information about a specific booking.
      operationId: getBooking
      tags:
        - Bookings
      parameters:
        - name: id
          in: path
          required: true
          description: Booking ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Booking"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      summary: Update booking status
      description: |
        Update the status of a booking. Some status changes trigger side effects.

        **Side Effects**:
        - `confirmed`: Sets confirmed_at timestamp, sends confirmation email
        - `cancelled`: Releases time slot, sets cancelled_at timestamp
      operationId: updateBookingStatus
      tags:
        - Bookings
      parameters:
        - name: id
          in: path
          required: true
          description: Booking ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: "#/components/schemas/BookingStatus"
                notes:
                  type: string
                  description: Optional notes about the status change
                  maxLength: 500
      responses:
        "200":
          description: Booking status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Booking"
...
  /quotes:
    post:
      summary: Generate a quote
      description: |
        Generate a detailed pricing quote based on service selections and guest count.

        **Business Logic**: Applies volume discounts, service tier multipliers, and add-on pricing.
      operationId: generateQuote
      tags:
        - Quotes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteRequest"
      responses:
        "201":
          description: Quote generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Quote"
                  breakdown:
                    $ref: "#/components/schemas/QuoteBreakdown"
  /quotes/{id}:
    get:
      summary: Get quote details
      operationId: getQuote
      tags:
        - Quotes
      parameters:
        - name: id
          in: path
          required: true
          description: Quote ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Quote details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Quote"

    patch:
      summary: Update quote status
      description: |
        Update the status of a quote.

        **Side Effects**:
        - `sent`: Sets sent_at timestamp, sends quote email
        - `accepted`: Sets accepted_at timestamp, may create booking
      operationId: updateQuoteStatus
      tags:
        - Quotes
      parameters:
        - name: id
          in: path
          required: true
          description: Quote ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: "#/components/schemas/QuoteStatus"
      responses:
        "200":
          description: Quote status updated successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "NOT_FOUND"
              message: "The requested resource was not found"
...
  schemas:
    ServiceCategory:
      type: string
      enum: [corporate, private, wedding, celebration]
      description: |
        Type of catering service:
        - `corporate`: Business events and corporate catering
        - `private`: Private parties and intimate gatherings  
        - `wedding`: Wedding ceremonies and receptions
        - `celebration`: Special celebrations and milestone events
    ServiceTier:
      type: string
      enum: [essential, premium, luxury]
      description: |
        Service level tiers:
        - `essential`: Basic service package with quality essentials
        - `premium`: Enhanced service with premium options
        - `luxury`: Top-tier service with luxury amenities
    BookingStatus:
      type: string
      enum: [pending, confirmed, cancelled, completed]
      description: |
        Booking status lifecycle:
        - `pending`: Initial booking, awaiting confirmation
        - `confirmed`: Booking confirmed and scheduled
        - `cancelled`: Booking cancelled (releases time slot)
        - `completed`: Event completed successfully
    QuoteStatus:
      type: string
      enum: [draft, sent, accepted, rejected, expired]
      description: |
        Quote status lifecycle:
        - `draft`: Quote created but not sent
        - `sent`: Quote sent to customer
        - `accepted`: Customer accepted the quote
        - `rejected`: Customer rejected the quote  
        - `expired`: Quote validity period expired
...
    Booking:
      type: object
      required:
        - id
        - customer_name
        - customer_email
        - event_date
        - event_time
        - guest_count
        - service_category
        - service_tier
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        customer_name:
          type: string
          maxLength: 100
          description: Full name of the customer
        customer_email:
          type: string
          format: email
          maxLength: 254
          description: Customer's email address
        customer_phone:
          type: string
          maxLength: 20
          nullable: true
          description: Customer's phone number
        company_name:
          type: string
          maxLength: 100
          nullable: true
          description: Company name for corporate events
        event_date:
          type: string
          format: date
          description: Date of the event
        event_time:
          type: string
          pattern: "^[0-2][0-9]:[0-5][0-9]$"
          description: Start time of the event
        guest_count:
          type: integer
          minimum: 10
          maximum: 500
          description: Number of guests expected
        service_category:
          $ref: "#/components/schemas/ServiceCategory"
        service_tier:
          $ref: "#/components/schemas/ServiceTier"
        status:
          $ref: "#/components/schemas/BookingStatus"
        special_requests:
          type: string
          maxLength: 1000
          nullable: true
          description: Any special requests or notes
        dietary_restrictions:
          type: string
          maxLength: 500
          nullable: true
          description: Dietary restrictions or allergies
        estimated_total:
          type: number
          format: float
          minimum: 0
          nullable: true
          description: Estimated total cost in EUR
        final_total:
          type: number
          format: float
          minimum: 0
          nullable: true
          description: Final confirmed total cost in EUR
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        confirmed_at:
          type: string
          format: date-time
          nullable: true
        cancelled_at:
          type: string
          format: date-time
          nullable: true
...
    Quote:
      type: object
      required:
        - id
        - service_details
        - pricing_breakdown
        - total_amount
        - status
        - valid_until
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        booking_id:
          type: string
          format: uuid
          nullable: true
          description: Associated booking ID if quote was accepted
        service_details:
          type: object
          description: Detailed breakdown of requested services
        pricing_breakdown:
          type: object
          description: Detailed pricing calculation
        total_amount:
          type: number
          format: float
          minimum: 0
          description: Total quote amount in EUR
        status:
          $ref: "#/components/schemas/QuoteStatus"
        valid_until:
          type: string
          format: date
          description: Quote expiration date
        selected_add_ons:
          type: array
          items:
            type: string
          description: Selected add-on service IDs
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        sent_at:
          type: string
          format: date-time
          nullable: true
        accepted_at:
          type: string
          format: date-time
          nullable: true
    CreateQuoteRequest:
      type: object
      required:
        - service_category
        - service_tier
        - guest_count
      properties:
        booking_id:
          type: string
          format: uuid
          description: Associate quote with existing booking
        service_category:
          $ref: "#/components/schemas/ServiceCategory"
        service_tier:
          $ref: "#/components/schemas/ServiceTier"
        guest_count:
          type: integer
          minimum: 10
          maximum: 500
        event_date:
          type: string
          format: date
          description: Event date for seasonal pricing
        selected_add_ons:
          type: array
          items:
            type: string
            format: uuid
          description: Add-on service IDs to include
        custom_notes:
          type: string
          maxLength: 500
          description: Special requirements or notes
    QuoteBreakdown:
      type: object
      properties:
        base_price:
          type: number
          format: float
          description: Base service price
        base_price_per_person:
          type: number
          format: float
          description: Base price per guest
        tier_multiplier:
          type: number
          format: float
          description: Service tier pricing multiplier
        subtotal:
          type: number
          format: float
          description: Subtotal before discounts
        add_ons:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              price_per_person:
                type: number
                format: float
                nullable: true
              flat_rate:
                type: number
                format: float
                nullable: true
              total:
                type: number
                format: float
        add_ons_total:
          type: number
          format: float
          description: Total cost of all add-ons
        volume_discount:
          type: object
          nullable: true
          properties:
            min_guests:
              type: integer
            discount:
              type: number
              format: float
            discount_amount:
              type: number
              format: float
            label:
              type: string
        final_total:
          type: number
          format: float
          description: Final total after all discounts
        price_per_person:
          type: number
          format: float
          description: Final price per guest
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
              additionalProperties: true
...
security:
  - BearerAuth: []
...